requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
docutils==0.18.1
gcloud==0.18.3
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
jws==0.1.3
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.2
mysql-connector==2.2.9
mysql-connector-python==8.0.27
oauth2client==4.1.3
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.12.0
Pygments==2.11.1
pyparsing==3.0.6
pypiwin32==223
Pyrebase4==4.5.0
python-jwt==2.0.1
pywin32==303
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
urllib3==1.26.7
